requirements.txt
streamlit
langchain-groq
langchain
requests
python-dotenv